Changing for CMake. Need to debug.
